requirements.txt
transformers
diffusers
scipy
git+https://github.com/openai/CLIP.git
matplotlib
pandas
fitter
seaborn
pytorch_lightning